Created a superclass for SPARQL query speed comparisons.
